feat: better separation of concerns
